Load company-org-roam backend with +roam flag.
The backend used to be set in `company-org-roam` package clause, but the
config is evaluated before the `set-company-backend!` call in `org`
package and is being overriden.

I could not find a way to postpone the config call after org package is
done, so I added the feature detection and decision directly to org
package's config section.